requirements.txt
discord.py
python-dotenv
flask
pytest
coverage
pytest-cov